split script into multiple files and add parent directory
