+++ packages/create-plugin/src/codemods/additions/additions.ts
@@ -9,6 +9,6 @@ export default [
   {
     name: 'i18n',
     description: 'Adds internationalization (i18n) support to the plugin',
-    scriptPath: import.meta.resolve('./scripts/i18n.js'),
+    scriptPath: import.meta.resolve('./scripts/i18n/index.js'),
   },
 ] satisfies Codemod[];
