+++ packages/create-plugin/src/codemods/additions/scripts/i18n/index.ts
@@ -1,0 +1,81 @@
+import * as v from 'valibot';
+
+import type { Context } from '../../../context.js';
+import { additionsDebug } from '../../../utils.js';
+import { updateDockerCompose, updatePluginJson, createI18nextConfig } from './config-updates.js';
+import { addI18nInitialization, createLoadResourcesFile } from './code-generation.js';
+import { updateEslintConfig, addI18nDependency, addSemverDependency, addI18nextCli } from './tooling.js';
+import { checkNeedsBackwardCompatibility, createLocaleFiles } from './utils.js';
+
+/**
+ * I18n addition schema using Valibot
+ * Adds internationalization support to a plugin
+ */
+export const schema = v.object(
+  {
+    locales: v.pipe(
+      v.union([v.string(), v.array(v.string())]),
+      v.transform((input) => {
+        // Handle both string (from CLI) and array (from tests)
+        return typeof input === 'string' ? input.split(',').map((s) => s.trim()) : input;
+      }),
+      v.array(
+        v.pipe(
+          v.string(),
+          v.regex(/^[a-z]{2}-[A-Z]{2}$/, 'Locale must be in format xx-XX (e.g., en-US, es-ES, sv-SE)')
+        ),
+        'Please provide a comma-separated list of all supported locales, e.g., "en-US,es-ES,sv-SE"'
+      ),
+      v.minLength(1, 'Please provide a comma-separated list of all supported locales, e.g., "en-US,es-ES,sv-SE"')
+    ),
+  },
+  'Please provide a comma-separated list of all supported locales, e.g., "en-US,es-ES,sv-SE"'
+);
+
+type I18nOptions = v.InferOutput<typeof schema>;
+
+export default function i18nAddition(context: Context, options: I18nOptions): Context {
+  const { locales } = options;
+
+  additionsDebug('Adding i18n support with locales:', locales);
+
+  // Determine if we need backward compatibility (Grafana < 12.1.0)
+  const needsBackwardCompatibility = checkNeedsBackwardCompatibility(context);
+  additionsDebug('Needs backward compatibility:', needsBackwardCompatibility);
+
+  // 1. Update docker-compose.yaml with feature toggle
+  updateDockerCompose(context);
+
+  // 2. Update plugin.json with languages and grafanaDependency
+  updatePluginJson(context, locales, needsBackwardCompatibility);
+
+  // 3. Create locale folders and files
+  createLocaleFiles(context, locales);
+
+  // 4. Add @grafana/i18n dependency
+  addI18nDependency(context);
+
+  // 5. Add semver dependency for backward compatibility
+  if (needsBackwardCompatibility) {
+    addSemverDependency(context);
+  }
+
+  // 6. Update eslint.config.mjs if needed
+  updateEslintConfig(context);
+
+  // 7. Add i18n initialization to module file
+  addI18nInitialization(context, needsBackwardCompatibility);
+
+  // 8. Create loadResources.ts for backward compatibility
+  if (needsBackwardCompatibility) {
+    createLoadResourcesFile(context);
+  }
+
+  // 9. Add i18next-cli as dev dependency and add script
+  addI18nextCli(context);
+
+  // 10. Create i18next.config.ts
+  createI18nextConfig(context);
+
+  return context;
+}
